requirements.txt
streamlit==1.28.1
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
langchain==0.1.0
langchain-groq==0.0.1
crewai==0.28.8
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
